Add new voice.properties property sip.ignoreCallStateCheck = True
This turns off the block of code that checks to see if the call state is not Connected. The call state check was something I recently added, and now I worry that it may be causing some false disconnects. I think I would like to rely on GCEV events for hangup detection instead. This new property allows me to test that